Moved DB tests to separate file. Moved common functions to separate file. Changed Item float to int for price. General refactor

+++ PharmacySystem/Item.cpp
@@ -24,7 +24,7 @@ public:
 
 	string getDescription() { return description; }
 
-	float getPrice() { return price; }
+	int getPrice() { return price; }
 
 	string getDosage() { return dosage; }
